I am working on getting the final test to pass. This has required adding some temporary console messages in the session loading logic. This will be removed later.
Why studying the tests to get a better idea how the session loading is supposed to work, I noticed some tests that could be improved. Since I still saw value in the original tests, I have left them and created copies with the improvements. I was not sure what to name the improvements, however. So I stuck a cute message at the end of the copied descriptions.